refactor: extract six feature hooks from tab-query-editor
tab-query-editor.tsx was 1697 lines mixing query execution with every
side feature. Extract each cohesive cluster into a hook under
query-editor/, preserving behaviour exactly:

- use-fk-panels: FK panel stack state + fetch/open/close handlers
- use-column-stats-panel: column stats store wiring + header click
- use-query-analysis: telemetry/benchmark/EXPLAIN/perf-analysis state
  and handlers (wraps useTabTelemetry + useTabPerfIndicator)
- use-manual-transaction: auto-commit toggle, commit/rollback, unmount
  and connection-change rollback safety nets
- use-result-export: current-result resolution, masked-export gating,
  clipboard/file export
- use-watch-runner: Watch Mode runner + menu-accelerator toggle

Component drops to 1144 lines of layout/orchestration. No behaviour
change intended; typecheck, 973 unit tests, and the full 46-test e2e
suite (transactions, watch, time machine) pass.
